feat(wasm): added FilePlayer
